split enumerator.ts from generator.ts and split enum test from gen test
